coverage module test improved, and passing
